Move some more stuff
* Renamed ActionIdentifier.ts to Actions/utils.ts
* Moved some more Action utility functions that don't need access to bladeburner over to utils.ts
* Change getActionIdFromTypeAndName to getActionFromTypeAndName - it was always immediately being followed up by getting the action from the actionId so the actionId was just a pointless middle step
* Startbutton now has the action instead of the actionId
