+++ src/PersonObjects/Sleeve/Work/SleeveBladeburnerWork.ts
@@ -8,7 +8,7 @@ import { applySleeveGains, SleeveWorkClass, SleeveWorkType } from "./Work";
 import { CONSTANTS } from "../../../Constants";
 import { scaleWorkStats } from "../../../Work/WorkStats";
 import { getKeyList } from "../../../utils/helpers/getKeyList";
-import { loadActionIdentifier } from "../../../Bladeburner/Actions/ActionIdentifier";
+import { getActionObject, loadActionIdentifier } from "../../../Bladeburner/Actions/utils";
 import { invalidWork } from "../../../Work/InvalidWork";
 
 interface SleeveBladeburnerWorkParams {
@@ -36,7 +36,7 @@ export class SleeveBladeburnerWork extends SleeveWorkClass {
 
   cyclesNeeded(sleeve: Sleeve): number {
     if (!Player.bladeburner) return Infinity;
-    const action = Player.bladeburner.getActionObject(this.actionId);
+    const action = getActionObject(this.actionId);
     const timeInMs = action.getActionTime(Player.bladeburner, sleeve) * 1000;
     return timeInMs / CONSTANTS.MilliPerCycle;
   }
@@ -49,13 +49,13 @@ export class SleeveBladeburnerWork extends SleeveWorkClass {
     if (!Player.bladeburner) return sleeve.stopWork();
     this.cyclesWorked += cycles;
     if (this.actionId.type === BladeActionType.contract) {
-      const action = Player.bladeburner.getActionObject(this.actionId);
+      const action = getActionObject(this.actionId);
       if (action.count < 1) return sleeve.stopWork();
     }
 
     while (this.cyclesWorked >= this.cyclesNeeded(sleeve)) {
       if (this.actionId.type === BladeActionType.contract) {
-        const action = Player.bladeburner.getActionObject(this.actionId);
+        const action = getActionObject(this.actionId);
         if (action.count < 1) return sleeve.stopWork();
       }
       const retValue = Player.bladeburner.completeAction(sleeve, this.actionId, false);
